requirements.txt
google_api_python_client==2.78.0
ipython==8.10.0
pandas==1.5.2
protobuf==4.22.0
pytest==7.2.0
PyYAML==6.0
requests==2.28.1
setuptools==65.5.0
zstd==1.5.2.6
